Changes for model helper
